refactor: promote bilevel solver to first-class mlsynth.utils.bilevel
The bilevel SCM solver (predictor-weight V + donor-weight W program) and its
high-level BilevelSCM engine were nested under fscm_helpers/ and
vanillasc_helpers/ respectively, forcing MASC, SpillSynth and others to import
shared infrastructure *through* sibling estimators. Promote the whole stack to
a first-class mlsynth/utils/bilevel/ package:

* git mv utils/fscm_helpers/bilevel -> utils/bilevel
* git mv utils/vanillasc_helpers/engine.py -> utils/bilevel/engine.py
  (BilevelSCM/BilevelSCMResult now exported from utils.bilevel)
* repoint all importers (fscm, masc, spillsynth, vanillasc) at utils.bilevel
* fix relative-import depth in the moved package

Pure move, no behavior change. Regression: test_bilevel_robustness,
test_determine_v, test_fscm_bilevel, test_vanillasc, test_fscm, test_masc,
test_spillsynth all green (261 passed); full suite collects clean (2944).
