+++ mlsynth/utils/bilevel/structure.py
@@ -20,7 +20,7 @@ from typing import Any, Dict, List
 
 import numpy as np
 
-from ....exceptions import MlsynthDataError
+from ...exceptions import MlsynthDataError
 
 
 @dataclass(frozen=True)
