+++ mlsynth/utils/spillsynth_helpers/iscm/weights.py
@@ -18,7 +18,7 @@ from typing import List, Optional, Tuple
 
 import numpy as np
 
-from ...fscm_helpers.bilevel import (
+from ...bilevel import (
     BilevelProblem, bias_corrected_gaps, simplex_lstsq, solve_bilevel)
