+++ mlsynth/utils/vanillasc_helpers/__init__.py
@@ -1,6 +1,6 @@
 """Helpers for the VanillaSC estimator (standard SCM on the bilevel engine)."""
 
-from .engine import BilevelSCM, BilevelSCMResult
+from ..bilevel import BilevelSCM, BilevelSCMResult
 from .lto import lto_placebo_test
 
 __all__ = ["BilevelSCM", "BilevelSCMResult", "lto_placebo_test"]
